% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astidytabulardata.R
\name{AsTidyTabularData}
\alias{AsTidyTabularData}
\title{Coerce an R Object To a Tidy Table}
\usage{
AsTidyTabularData(x)
}
\arguments{
\item{x}{An object to be coerced to a tidy table.}
}
\value{
A \strong{named} matrix or
    vector; a tidy version of \code{x}
}
\description{
Creates tables used by Displayr charting functions
}
\details{
\code{factors} will automatically be converted to numeric
    variables (with a warning) using
    \code{\link[flipTransformations]{AsNumeric}}.
}
\seealso{
\code{\link[flipTransformations]{AsNumeric}}
}
