% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{AutoOrderColumns}
\alias{AutoOrderColumns}
\title{Automatically order the columns by correspondence analysis of the table}
\usage{
AutoOrderColumns(x)
}
\arguments{
\item{x}{Input matrix}
}
\description{
Columns of the table are ordered according to the column-coordinates
   given by correspondence analysis of the table (1st dimension)
}
