% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{AutoOrderRows}
\alias{AutoOrderRows}
\title{Automatically order the rows by correspondence analysis of the table}
\usage{
AutoOrderRows(x)
}
\arguments{
\item{x}{Input matrix}
}
\description{
Rows of the table are ordered according to the row-coordinates
   given by correspondence analysis of the table (1st dimension)
}
