% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTable.R
\name{BasicTable}
\alias{BasicTable}
\title{Create a BasicTable Object}
\usage{
BasicTable(x, date = NULL, row.names.to.remove = NULL,
  col.names.to.remove = NULL, transpose = FALSE)
}
\arguments{
\item{x}{a \strong{named} numeric matrix or vector}

\item{date}{optional vector of dates, which if supplied, will be
used for the row names of the returned table; must be unique
and have the same length or number of rows of \code{x}; will
overwrite any existing row names present in \code{x}}

\item{row.names.to.remove}{character vector of row labels
specifying rows to remove from the returned table}

\item{col.names.to.remove}{character vector of column labels
specifying columns to remove from the returned table}

\item{transpose}{logical; if \code{TRUE} the table will be
transposed before being returned}
}
\value{
An object of class \code{BasicTable} - a \strong{named}
    matrix or vector
}
\description{
Creates tables used by Displayr charting functions
}
\details{
If \code{x} is not a numeric vector or matrix, an attempt
    will be made to coerce it to one using \code{\link{AsBasicTable}}.
}
\seealso{
\code{\link{AsBasicTable}}
}
