% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_and_c_bind.R
\name{Cbind}
\alias{Cbind}
\title{Cbind}
\usage{
Cbind(..., keep.all = TRUE)
}
\arguments{
\item{...}{The tables or vectors to be merged.}

\item{keep.all}{If \code{"TRUE"}, even non-matching rows are retained. Ignored and treated as
\code{"TRUE"} if matching is based upon indices.}
}
\description{
Takes a sequence of vectors or tables and merges them, similar to \code{\link{cbind}}, except that it matches
based on names. If there are no names, matching is based upon indices.
}
