% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertqtabletoarray.R
\name{ConvertQTableToArray}
\alias{ConvertQTableToArray}
\title{Converts a Q Table into an array}
\usage{
ConvertQTableToArray(x)
}
\arguments{
\item{x}{A Q table which is expected to have the attributes
typically associated with tables, e.g. "questions", "name"
and "statistic" (if there is only one statistic).}
}
\value{
The output is always a 3d array, so that 1-column vector
is converted to an array with 3 dimensions (but last two have
a length of 1). This makes it easy to identify the statistics.
}
\description{
Converts a Q Table into an array
}
