% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiderowsandcolumns.R
\name{GetNonEmptyElements}
\alias{GetNonEmptyElements}
\title{Get Names of Empty Entries From a Table Or Vector}
\usage{
GetNonEmptyElements(x, use.names = TRUE, is.percent = NULL)
}
\arguments{
\item{x}{matrix; vector; or list}

\item{use.names}{logical; if \code{TRUE}, the nams of the non-empty
entries are returned; otherwise, the numeric positions of the
non-empty elements are returned}

\item{is.percent}{logical; indicating whether the supplied data are
percentages; specifying a value of \code{NULL} (the default)
means that the function will check for an attribute of \code{x}
called "statistic" to check for percentages}
}
\value{
a vector containing the non-empty indices
}
\description{
Searches for empty rows or columns in a table and returns a copy
with those entries removed
}
