% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiderowsandcolumns.R
\name{GetNonEmptyRowsAndColumns}
\alias{GetNonEmptyRowsAndColumns}
\title{Get Names of Empty Entries From a Table Or Vector}
\usage{
GetNonEmptyRowsAndColumns(x, use.names = TRUE, is.percent = NULL)
}
\arguments{
\item{x}{matrix; vector; or list}

\item{use.names}{logical; if \code{TRUE}}

\item{is.percent}{logical; indicating whether the supplied data are
percentages; specifying a value of \code{NULL} (the default)
means that the function will check for an attribute of \code{x}
called "statistic" to check for percentages}
}
\value{
Two-element list containing the non-empty indices
}
\description{
Searches for empty rows or columns in a table and returns a copy
with those entries removed
}
