% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hideemptyrowsandcolumns.R
\name{HideEmptyColumns}
\alias{HideEmptyColumns}
\title{Hide Empty Columns From a Table}
\usage{
HideEmptyColumns(x, remove.zeros = TRUE)
}
\arguments{
\item{x}{matrix; vector; or list}

\item{remove.zeros}{logical; indicating whether zeros are removed.
If \code{FALSE}, then only \code{NAs} are removed.
Specifying a value of \code{NULL}
means that the function will check for an attribute of \code{x}
called "statistic" to check for percentages}
}
\description{
Searches for empty columns in a table and returns a copy
with those entries removed
}
