% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hideemptyrowsandcolumns.R
\name{HideEmptyRows}
\alias{HideEmptyRows}
\title{Hide Empty Rows From a Table}
\usage{
HideEmptyRows(x, is.percent = NULL)
}
\arguments{
\item{x}{matrix; vector; or list}

\item{is.percent}{logical; indicating whether the supplied data are
percentages; specifying a value of \code{NULL} (the default)
means that the function will check for an attribute of \code{x}
called "statistic" to check for percentages}
}
\value{
\code{x} with any empty rows or columns removed; or in the
    case \code{x} is vector, \code{x} with any empty entries
    removed
}
\description{
Searches for empty rows in a table and returns a copy
with those entries removed
}
\seealso{
\code{GetNonEmptyRowsAndColumns}
}
