% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hideemptyrowsandcolumns.R
\name{HideEmptyRowsAndColumns}
\alias{HideEmptyRowsAndColumns}
\title{Hide Empty Entries From a Table Or Vector}
\usage{
HideEmptyRowsAndColumns(x, is.percent = NULL)
}
\arguments{
\item{x}{matrix; vector; or list}

\item{is.percent}{logical; indicating whether the supplied data are
percentages; specifying a value of \code{NULL} (the default)
means that the function will check for an attribute of \code{x}
called "statistic" to check for percentages}
}
\value{
\code{x} with any empty rows or columns removed; or in the
    case \code{x} is vector, \code{x} with any empty entries
    removed
}
\description{
Searches for empty rows or columns in a table and returns a copy
with those entries removed
}
\examples{
x <- matrix(nrow = 2, ncol = 2)
x[1, ] <- 1
HideEmptyRowsAndColumns(x)
x[2, ] <- 0
HideEmptyRowsAndColumns(x, is.percent = TRUE)
}
\seealso{
\code{GetNonEmptyRowsAndColumns}
}
