% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{HideRowsWithSmallSampleSizes}
\alias{HideRowsWithSmallSampleSizes}
\title{Removes rows with small sample sizes}
\usage{
HideRowsWithSmallSampleSizes(x, min.size = 30)
}
\arguments{
\item{x}{Input table, which must contain the 'Column n' or 'Base n' statistic.}

\item{min.size}{Minimum sample size required.}
}
\description{
Using the 'Column n' (preferred) or 'Base n' statistic
  this funcion will remove any rows from \code{x}, where all
  are smaller than \code{min.size}. If any rows/columns are removed
  then warnings will be given.
}
