% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{HideValuesWithSmallSampleSizes}
\alias{HideValuesWithSmallSampleSizes}
\title{Removes values with small sample sizes}
\usage{
HideValuesWithSmallSampleSizes(x, min.size = 30)
}
\arguments{
\item{x}{Input table, which must contain the 'Column n' or 'Base n' statistic.}

\item{min.size}{Minimum sample size required.}
}
\description{
Using the 'Column n' (preferred) or 'Base n' statistic
  this funcion will remove any values from \code{x}, where all
  are smaller than \code{min.size}. They will be replaced with \code{NA}.
}
