% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{LastKColumns}
\alias{LastKColumns}
\title{LastKColumns}
\usage{
LastKColumns(x, K)
}
\arguments{
\item{x}{Input matrix or dataframe}

\item{K}{Number of columns to retain.}
}
\description{
Extract last K columns from table
}
