% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{LastKRows}
\alias{LastKRows}
\title{LastKRows}
\usage{
LastKRows(x, K)
}
\arguments{
\item{x}{Input matrix or dataframe}

\item{K}{Number of rows to retain.}
}
\description{
Extract last K rows from table
}
