% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergetables.R
\name{MergeTables}
\alias{Merge2Tables}
\alias{MergeTables}
\title{Merge tables}
\usage{
MergeTables(tables, joinby = c("Join columns", "Join rows"),
  nonmatching = c("Matching only", "Keep all"))

Merge2Tables(left, right, joinby = c("Join columns", "Join rows"),
  nonmatching = c("Matching only", "Keep all from first table",
  "Keep all from second table", "Keep all"))
}
\arguments{
\item{tables}{A list of tables to merge}

\item{joinby}{\code{"Join columns"} is similar to \code{\link{cbind}}.
\code{"Join rows"} is similar to \code{\link{rbind}}.}

\item{nonmatching}{How to handle non-matching row or column names. These are
similar to the \code{all.*} arguments in \code{\link{merge}}.
\code{MergeTables} supports 2 options: \code{"Matching only"} (like
\code{all = FALSE}) and \code{"Keep all"} (like \code{all = TRUE}).
\code{Merge2Tables} supports these and a further 2 options: \code{"Keep all
from first table"} (like \code{all.x = TRUE}) and \code{"Keep all from
second table"} (like \code{all.y = TRUE}).}

\item{left, right}{The tables to merge. These should be vectors, matrices or
arrays. If the array has 3 dimensions, the first 'plane' of the third
dimension is kept, the others are dropped. It is an error to have more than
3 dimensions in the array.}
}
\description{
Bind tables together either by rows (similar to \code{\link{rbind}}) or by
columns (similar to \code{\link{cbind}}). Unlike those functions, this will
attempt to match cases based on the row or column names (similar to
\code{\link{merge}}).
}
\section{Functions}{
\itemize{
\item \code{Merge2Tables}: Merge 2 tables.
}}

