% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padNAs.R
\name{padNAs}
\alias{padNAs}
\title{Pad with missing values}
\usage{
padNAs(x, add.n)
}
\arguments{
\item{x}{object to pad}

\item{add.n}{Number of empty entries or rows to add}
}
\description{
Add entries to vectors or rows to matrices
     or data frames filled with NAs.
}
