% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_and_c_bind.R
\name{Rbind}
\alias{Rbind}
\title{RBind}
\usage{
Rbind(...)
}
\arguments{
\item{...}{The tables or vectors to be merged.}
}
\description{
Takes a sequence of vectors or tables and merges them, similar to rbind, except that it matches
based on names.
}

