% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_and_c_bind.R
\name{rbindAndCbindWithLabels}
\alias{rbindAndCbindWithLabels}
\title{rbindAndCbindWithLabels}
\usage{
rbindAndCbindWithLabels(..., rows, keep.all)
}
\arguments{
\item{...}{The tables or vectors to be merged.}

\item{rows}{If TRUE, binds by rows. Otherwise by columns.}

\item{keep.all}{If TRUE, even non-matching rows are retained.}
}
\description{
Takes a sequence of vectors or tables and merges them, similar to rbind and cbind, except that it matches
based on names.
}
