% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{Reorder}
\alias{Reorder}
\title{Reorder}
\usage{
Reorder(x, rows = "Descending", columns = "Descending", FUN = mean,
  fudge = 1e-06)
}
\arguments{
\item{x}{A \code{\link{matrix}} or \code{\link{data.frame}} A list of tables to merge}

\item{rows}{One of  \code{"Descending"}, which orders the rows from top to bottom
based on \code{FUN}, \code{"Ascending"}, or \code{"None"}.}

\item{columns}{Same as \code{"rows"}.}

\item{FUN}{A function to be applied to the rows and/or columns as the
basis for the re-odering. Defaults to \code{mean}.}

\item{fudge}{The value of fudge * i, where i is the row or column number is
added to each cell. This ensure that ties are broken (provided that the values of
the data are not very small, and that they are broken in a consistent and likely
orderly way (e.g., if all the numbers in the table are the same, the table
is not reordered).}
}
\description{
Reorders the rows or columns of a table.
}

