% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{ReverseColumns}
\alias{ReverseColumns}
\title{Reverse columns of a table}
\usage{
ReverseColumns(x)
}
\arguments{
\item{x}{Input table which may be a matrix, dataframe or array.
Vectors or 1-d arrays will be ignored}
}
\description{
Reverse order of columns in matrix, dataframe or array
}
