% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{ReverseRows}
\alias{ReverseRows}
\title{Reverse rows of a table}
\usage{
ReverseRows(x)
}
\arguments{
\item{x}{Input table which may be a matrix, dataframe or array.
Vectors or 1-d arrays will be converted into a matrix with 1 column}
}
\description{
Reverse order of rows in matrix, dataframe or array
}
