% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{SortColumns}
\alias{SortColumns}
\title{Sort columns of a table}
\usage{
SortColumns(x, decreasing = FALSE, row = NULL,
  exclude = "NET, SUM, Total")
}
\arguments{
\item{x}{Input matrix or dataframe which is being sorted. Values must be numeric.
Vectors and 1-d arrays will be ignored.}

\item{decreasing}{Order to sort values.}

\item{row}{The row to sort by. If none is specified and the 'Row n'
statistic is present in the table, it will use the row with the largest
value of 'Row n'. Otherwise it will pick the bottom row.}

\item{exclude}{A string containing a comma-separated list of columns
(either by name or index) which should not be sorted. These columns
will remain at the end of the output table}
}
\description{
Sorts the columns of the table based on the values in the specified column
}
