% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytabulardata.R
\name{TidyTabularData}
\alias{TidyTabularData}
\title{Create a Tidy Table or Vector}
\usage{
TidyTabularData(x, date = FALSE, row.names.to.remove = NULL,
  col.names.to.remove = NULL, transpose = FALSE)
}
\arguments{
\item{x}{a numeric matrix or vector; should be named, but if names
are not present they will be created in the output}

\item{date}{optional vector of dates, which if supplied, will be
used for the row names of the returned table; must be unique
and have the same length or number of rows of \code{x}; will
overwrite any existing row names present in \code{x}; the
default, \code{FALSE}, means that no dates are present in the
data; a value of \code{NULL} indicates that dates are contained
in x}

\item{row.names.to.remove}{character vector of row labels
specifying rows to remove from the returned table}

\item{col.names.to.remove}{character vector of column labels
specifying columns to remove from the returned table}

\item{transpose}{logical; if \code{TRUE} the table will be
transposed before being returned}
}
\value{
A \strong{named} matrix or vector; a tidy version of \code{x}
}
\description{
Creates tables used by Displayr charting functions
}
\details{
If \code{x} is not a numeric vector or matrix, an attempt
    will be made to coerce it to one using
    \code{\link{AsTidyTabularData}}.

If a named vector \code{BasicTable} is created from \code{x}, then
\code{row.names.to.remove} and \code{col.names.to.remove} will be
combined (using \code{\link[base]{union}}) to determine entries to remove.
}
\note{
If \code{transpose == TRUE}, then the table is transposed
\emph{before} rows and columns are removed, so
\code{row.names.to.remove} should be specified according to the
rownames of the \emph{transposed} table, as should
\code{col.names.to.remove}
}
\seealso{
\code{\link{AsTidyTabularData}}
}
