% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{TopKAndSpecifiedRows}
\alias{TopKAndSpecifiedRows}
\title{\code{TopKAndSpecifiedRows}}
\usage{
TopKAndSpecifiedRows(x, K, select = NULL)
}
\arguments{
\item{x}{Input table}

\item{K}{Number of rows from the top of the table to return.}

\item{select}{Optional string containing comma separated list of names 
or indices of the rows to be selected}
}
\description{
Returns the first K rows (no sorting), and additional rows listed
}
