% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsBasicTable.R
\name{AsBasicTable}
\alias{AsBasicTable}
\title{Create a BasicTable Object}
\usage{
AsBasicTable(x)
}
\arguments{
\item{x}{\strong{named} matrix or vector}
}
\value{
An object of class \code{BasicTable} - a named matrix or vector
}
\description{
Creates tables used by Displayr charting functions
}
