% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsBasicTable.R
\name{AsBasicTable}
\alias{AsBasicTable}
\title{Create a BasicTable Object}
\usage{
AsBasicTable(x)
}
\arguments{
\item{x}{\strong{named} matrix or vector}
}
\value{
An object of class \code{BasicTable} - a named matrix or
    vector
}
\description{
Creates tables used by Displayr charting functions
}
\details{
\code{factors} will automatically be converted to numeric
    variables (with a warning) using
    \code{\link[flipTransformations]{AsNumeric}}; with behaviour
    dependent upon the value of \code{as.binary}
}
\seealso{
\code{\link[flipTransformations]{AsNumeric}}
}
