% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astidytabulardata.R
\name{AsTidyTabularData}
\alias{AsTidyTabularData}
\title{Coerce an R Object To a Tidy Table}
\usage{
AsTidyTabularData(x)
}
\arguments{
\item{x}{An object to be coerced to a tidy table.}
}
\value{
A \strong{named} matrix or
    vector; a tidy version of \code{x}
}
\description{
Coerces R objects to a tidy tabular format.  Recognized inputs include
tables from \code{Q}, arrays, matrices, data.frames, and pasted
tables (character matrices).  Output will always be named and always
be either a vector or matrix.
}
\details{
\code{factors} will automatically be converted to numeric
    variables (with a warning) using
    \code{\link[flipTransformations]{AsNumeric}}.
}
\seealso{
\code{\link[flipTransformations]{AsNumeric}}
}
