% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_and_c_bind.R
\name{Cbind}
\alias{Cbind}
\title{CBind}
\usage{
Cbind(..., keep.all = TRUE)
}
\arguments{
\item{...}{The tables or vectors to be merged.}

\item{keep.all}{If TRUE, even non-matching rows are retained.}
}
\description{
Takes a sequence of vectors or tables and merges them, similar to \code{\link{rbind}}, except that it matches
based on names.
}
