% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{HideOutputsWithSmallSampleSizes}
\alias{HideOutputsWithSmallSampleSizes}
\title{Throws error if table has small sample size}
\usage{
HideOutputsWithSmallSampleSizes(x, min.size = 30)
}
\arguments{
\item{x}{Input table, which must contain the 'Base n' statistic.}

\item{min.size}{Minimum sample size required.}
}
\description{
Throws an error if any of the 'Base n' values in the table are too small
}
