% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeTables.R
\name{MergeTables}
\alias{MergeTables}
\title{Merge tables}
\usage{
MergeTables(left, right, joinby = c("Join columns", "Join rows"),
  nonmatching = c("Matching only", "Keep all from first table",
  "Keep all from second table", "Keep all"))
}
\arguments{
\item{left, right}{The tables to merge. These should be vectors, matrices
or arrays. If the array has 3 dimensions, the first 'plane' of the
third dimension is kept, the others are dropped. It is an error to
have more than 3 dimensions in the array.}

\item{joinby}{'Join columns' is similar to \code{\link{cbind}}.
'Join rows' is similar to \code{\link{rbind}}.}

\item{nonmatching}{How to handle non-matching row or column names.
These are similar to the \code{all.*} arguments in
\code{\link{merge}}.}
}
\description{
Bind tables together either by rows (similar to \code{\link{rbind}})
or by columns (similar to \code{\link{cbind}}). Unlike those
functions, this will attempt to match cases based on the row or column
names (similar to \code{\link{merge}}).
}

