% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergetables.R
\name{MergeTables}
\alias{Merge2Tables}
\alias{MergeTables}
\title{MergeTables}
\usage{
MergeTables(tables, direction = c("Side-by-side", "Up-and-down"),
  nonmatching = c("Keep all", "Matching only"))

Merge2Tables(left, right, direction = c("Side-by-side", "Up-and-down"),
  nonmatching = c("Keep all", "Keep all from first table",
  "Keep all from second table", "Matching only"))
}
\arguments{
\item{tables}{A list of tables to merge}

\item{direction}{\code{"Side-by-side"} is similar to \code{\link{cbind}}.
\code{"Up-and-down"} is similar to \code{\link{rbind}}.}

\item{nonmatching}{How to handle non-matching row or column names. These are
similar to the \code{all.*} arguments in \code{\link{merge}}.
\code{MergeTables} supports 2 options: \code{"Keep all"} (like \code{all =
TRUE}) and \code{"Matching only"} (like \code{all = FALSE}).
\code{Merge2Tables} supports these and a further 2 options: \code{"Keep all
from first table"} (like \code{all.x = TRUE}) and \code{"Keep all from
second table"} (like \code{all.y = TRUE}).}

\item{left, right}{The tables to merge. These should be vectors, matrices or
arrays. If the array has 3 dimensions, the first 'plane' of the third
dimension is kept, the others are dropped. It is an error to have more than
3 dimensions in the array.}
}
\description{
Merge two or more tables. Bind tables together either by rows (similar to \code{\link{rbind}}) or by
columns (similar to \code{\link{cbind}}). Unlike those functions, this will
attempt to match cases based on the row or column names (similar to
\code{\link{merge}}).
}
\section{Functions}{
\itemize{
\item \code{Merge2Tables}: Merge two tables.
}}

