% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_and_c_bind.R
\name{rbindAndCbindWithLabels}
\alias{rbindAndCbindWithLabels}
\title{rbindAndCbindWithLabels}
\usage{
rbindAndCbindWithLabels(..., rows, keep.all)
}
\arguments{
\item{...}{The tables or vectors to be merged.}

\item{rows}{If TRUE, binds by rows. Otherwise by columns.}

\item{keep.all}{If \code{"TRUE"}, even non-matching rows/columns are retained. Ignored and treated as
\code{"TRUE"} if matching is based upon indices.}
}
\description{
Takes a sequence of vectors or tables and merges them, similar to rbind and cbind, except that it matches
based on names. If there are no names, matching is based upon indices.
}
