% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removerowsandorcolumns.R
\name{RemoveRowsAndOrColumns}
\alias{RemoveRowsAndOrColumns}
\title{\code{RemoveRowsAndOrColumns}}
\usage{
RemoveRowsAndOrColumns(x, row.names.to.remove = c("NET", "Total", "SUM"),
  column.names.to.remove = c("NET", "Total", "SUM"), split = "[;,]")
}
\arguments{
\item{x}{Data to remove rows and columns from; can be a matrix or data.frame or
a list of matrices and data.frames.}

\item{row.names.to.remove}{A vector or comma-separated string containing the row labels to remove.}

\item{column.names.to.remove}{A vector or comma-separated string containing the column labels to remove.}

\item{split}{Delimiter to split string on.}
}
\description{
Removes rows or columns from the table.
}
\details{
Trailing spaces are removed and lower/upper case is ignored.
}
