% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{SelectColumns}
\alias{SelectColumns}
\title{\code{SelectColumns}}
\usage{
SelectColumns(x, select = "")
}
\arguments{
\item{x}{Matrix or dataframe from which columns will be extracted}

\item{select}{A string containing a comma seperated list of the
names or indices of the columns to be selected.}
}
\description{
Select rows from a matrix or dataframe.
}
