% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{SelectEntry}
\alias{SelectEntry}
\title{Select entries from a table}
\usage{
SelectEntry(x, row, column = 1, return.single.value = FALSE)
}
\arguments{
\item{x}{Matrix or dataframe from which values will be extracted.}

\item{row}{A string containing a comma seperated list of the
name or indices of the rows to be selected}

\item{column}{A string containing a comma seperated list of the
name or indices of the column to be selected, If \code{x} is a vector
or 1-dimensional array, \code{column} will be ignored.}

\item{return.single.value}{Logical; If this true, the function will
always return a single numeric value. If multiple cells are selected,
only the first is returned. If no entries are selected a
value of zero is returned.}
}
\description{
Function to select entries from a table using
either the row/column name or indices
}
