% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytools.R
\name{SelectRows}
\alias{SelectRows}
\title{\code{SelectRows}}
\usage{
SelectRows(x, select = NULL)
}
\arguments{
\item{x}{Matrix or dataframe from which rows will be extracted}

\item{select}{A string containing a comma seperated list of the
names or indices of the rows to be selected.}
}
\description{
Select rows from a matrix or dataframe.
}
