% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytabulardata.R
\name{TidyTabularData}
\alias{TidyTabularData}
\title{Create a Tidy Table or Vector}
\usage{
TidyTabularData(
  x,
  date = FALSE,
  row.names.to.remove = NULL,
  col.names.to.remove = NULL,
  transpose = FALSE,
  split = "[;,]",
  hide.empty.rows.and.columns = FALSE,
  ...
)
}
\arguments{
\item{x}{Numeric matrix or vector; should be named, but if names
are not present they will be created in the output}

\item{date}{Optional vector of dates, which if supplied, will be
used for the row names of the returned table; must be unique
and have the same length or number of rows of \code{x}; will
overwrite any existing row names present in \code{x}; the
default, \code{FALSE}, means that no dates are present in the
data; a value of \code{NULL} indicates that dates are contained
in x.}

\item{row.names.to.remove}{Character vector or delimited string of
row labels specifying rows to remove from the returned table.}

\item{col.names.to.remove}{Character vector or delimited string of
column labels specifying columns to remove from the returned
table}

\item{transpose}{logical; if \code{TRUE} the table will be
transposed before being returned}

\item{split}{Character delimiter to split
\code{row.names.to.remove} and \code{col.names.to.remove}
on. Default is to split on either of \code{","} or \code{";"}.
Assumed to be a regular expression; see \code{\link{strsplit}}.}

\item{hide.empty.rows.and.columns}{Logical; if \code{TRUE} empty
rows and columns will be removed from the data.  Empty here
meaning that a row or column contains all \code{NA} values, or
in the case of percentages, that a row or column contains only
0's.}

\item{...}{Additional arguments passed to
\code{\link[flipTransformations]{ParseUserEnteredTable}} in the
case that \code{x} has character entries.}
}
\value{
A \strong{named} matrix or vector; a tidy version of \code{x}.
}
\description{
Creates tables that are ensured to be in a tidy format.  Output is
guaranteed to always have row and column names and there is special
handling of arrays and tables from \code{Q}.
}
\details{
If \code{x} is not a numeric vector or matrix, an attempt
    will be made to coerce it to one using
    \code{\link{AsTidyTabularData}}.

If a named vector is created from \code{x}, then
\code{row.names.to.remove} and \code{col.names.to.remove} will be
combined (using \code{\link[base]{union}}) to determine entries to remove.
}
\note{
If \code{transpose == TRUE}, then the table is transposed
\emph{before} rows and columns are removed, so
\code{row.names.to.remove} should be specified according to the
rownames of the \emph{transposed} table, as should
\code{col.names.to.remove}
}
\seealso{
\code{\link{AsTidyTabularData}}, \code{\link{HideEmptyRowsAndColumns}},
\code{\link{RemoveRowsAndOrColumns}}
}
